+++ core-simple/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>new_core</artifactId>
+    <artifactId>core-simple</artifactId>
 
     <properties>
         <maven.compiler.source>14</maven.compiler.source>
