+++ core-simple/src/main/java/com/example/demo/domain/StockEntity.java
@@ -1,28 +1,40 @@
 package com.example.demo.domain;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
-@Table(name = "stock")
+@Table(name = "stock",
+        uniqueConstraints = {
+                @UniqueConstraint(name = "shoe", columnNames = {"name", "size", "color"})
+        }
+)
 @Getter
 @Setter
+@NoArgsConstructor
 public class StockEntity {
-
-    private static final int MAX_CAPACITY = 30;
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "stock_generator")
     @SequenceGenerator(name = "stock_generator", sequenceName = "stock_id_seq", allocationSize = 1)
     private long id;
 
-    @ManyToOne
-    private ShoeEntity shoe;
+    private String name;
 
+    private BigInteger size;
+
+    private Color color;
 
     private Integer quantity;
 
+    public enum Color {
+        BLACK,
+        BLUE,
+        ;
+    }
 
 }
