+++ core-simple/src/main/resources/application.properties
